// src/app/models/movimiento.ts
export interface Movimiento{
    nombre_moneda_original: string
    valor_moneda_original: string
    cantidad_moneda_original: string
    nombre_moneda_nueva: string
    valor_moneda_nueva: string
    cantidad_moneda_nueva: string
    fecha_operacion: string
    usuario: string
    createdAt?: string
    updatedAt?: string
    _id?: string
}